backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.0
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
pydantic-settings==2.5.2
